% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateKeggAtlasLegend.R
\name{generateKeggAtlasHeatmapLegend}
\alias{generateKeggAtlasHeatmapLegend}
\title{Make a legend for the KEGG Atlas heatmap}
\usage{
generateKeggAtlasHeatmapLegend(
  compound_legend = "none",
  compound_legend_title = "Compounds",
  compound_legend_title_cex = 2,
  compound_colors = c("#00cc33", "#0018A9", "#333333"),
  compound_labels = c("Upregulated", "Downregulated", "Conflicting"),
  compound_labels_cex = 1.5,
  pathway_legend = "none",
  pathway_legend_title = "Pathways",
  pathway_legend_title_cex = 2,
  pathway_colors = colorRamp2(c(1:7), brewer.pal(7, "YlOrRd")),
  pathway_labels = c(1, 3, 5, 7),
  pathway_labels_cex = 1.5
)
}
\arguments{
\item{compound_legend}{Draws a legend for compounds. Can be continuous or discrete.}

\item{compound_legend_title}{Legend title as string.}

\item{compound_colors}{For discrete, this takes a vector list of colors. For continuous, this takes a colorRamp2 object (e.g. colorRamp2(c(1:7), brewer.pal(7, "YlOrRd")))}

\item{compound_labels}{For discrete, this takes a vector of the same length and order as the colors. For continuous, this takes a vector of the labels you want.}

\item{pathway_legend}{Draws a legend for pathways. Can be continuous or discrete.}

\item{pathway_legend_title}{Legend title as string}

\item{pathway_colors}{For discrete, this takes a vector list of colors. For continuous, this takes a colorRamp2 object (e.g. colorRamp2(c(1:7), brewer.pal(7, "YlOrRd")))}

\item{pathway_labels}{For discrete, this takes a vector of the same length and order as the colors. For continuous, this takes a vector of the labels you want.}
}
\description{
Draws a legend that you can add to the figure. If the legend is getting clipped, expand your window.
}
\examples{
generateKeggAtlasHeatmapLegend(pathway_legend = "continuous", compound_legend = "discrete")

generateKeggAtlasHeatmapLegend(
  pathway_legend = "discrete",
  pathway_colors = c("red", "blue"),
  pathway_labels = c("Up", "Down"),
  compound_legend = "continuous",
  compound_colors = colorRamp2(c(1:7), brewer.pal(7, "YlOrRd")),
  compound_labels = c(1, 4, 7)
)
}
