% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKeggAtlas.R
\name{getKeggAtlas}
\alias{getKeggAtlas}
\title{Generate a Kegg Atlas Heatmap}
\usage{
getKeggAtlas(
  df,
  download_file_path = getwd(),
  highlight_path_names = FALSE,
  module_names_font_color = "#ffffff",
  module_names_font_weight = "bold",
  module_names_background_color = "Default",
  module_names_background_stroke = "Default"
)
}
\arguments{
\item{df}{A data frame with 1-4 columns. The first column should be 'KEGG.ID' (can be compounds "C00001" or paths "map00012"), then 'Color' in hex form, then 'Line.Width' in the from "W12", then 'Opacity' between 0 and 1.}

\item{highlight_path_names}{Boolean to bold the highlighted KEGG pathways.}

\item{module_names_font_color}{Color of the module labels (i.e. Carbohydrate metabolism, Vitamins and co-factors).}

\item{module_names_font_weight}{Weight of the module labels. Options are 'bold', 'normal', 'bolder', 'lighter', or a numeric between 1 and 1000.}

\item{module_names_background_color}{Change the background colors of the module labels. iPath defaults to colors of the rainbow. This allows you to insert a color in hex or color names (e.g. "grey") to change all the module fill colors at the same time.}

\item{module_names_background_stroke}{This allows you to put a boarder around the module label background. Default is no stroke.}
}
\description{
This function runs a headless browser using 'chromote' and opens the iPath3 metabolic pathways website: 'https://pathways.embl.de/ipath3.cgi?map=metabolic'.
It will submit your data frame with pathways or compounds and allow you to do minor customizations.
It will then download a png of your data to a specified path name.
IMPORTANT: This function requires internet access and requires that you have Google Chrome installed.
}
\examples{
df <- data.frame(
  ID = c("C00025", "C00022"),
  Color = c("#00cc33", "#ff3333"),
  Width = c("W14", "W10")
)
getKeggAtlas(
  df,
  download_file_path = getwd(),
  highlight_path_names = TRUE,
  module_names_font_color = "black",
  module_names_font_weight = "bolder",
  module_names_background_color = "#dddddd",
  module_names_background_stroke = "black"
 )

}
