% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKeggAtlasFromMetaboAnalystOutput.R
\name{getKeggAtlasFromMetaboAnalystOutput}
\alias{getKeggAtlasFromMetaboAnalystOutput}
\title{Make a Kegg Atlas Heatmap from a MetaboAnalyst directory}
\usage{
getKeggAtlasFromMetaboAnalystOutput(
  metaboanalyst_directory = "C:/Users/grarat/Downloads/Download",
  paths_color_by = "Enrichment",
  paths_color_pallette = brewer.pal(7, "YlOrRd"),
  paths_significant_only = "all",
  paths_significant_cutoff = 0.05,
  paths_weight = "W10",
  paths_opacity = 1,
  include_compounds = "significant",
  compounds_color = "#00ff00",
  compounds_weight = "W10",
  compounds_opacity = 1,
  download_file_path = getwd(),
  download_file_name = "RiPath3 Kegg Atlas.png",
  highlight_path_names = TRUE,
  module_names_font_color = "#ffffff",
  module_names_font_weight = "bold",
  module_names_background_color = "Default",
  module_names_background_stroke = "Default",
  generate_legend = TRUE
)
}
\arguments{
\item{metaboanalyst_directory}{When you run MetaboAnalyst, download the entire zip folder and put the path here}

\item{paths_color_by}{Select the output you would like to color by.}

\item{paths_color_pallette}{Set color palette from RColorBrewer. For example: brewer.pal(7, "YlOrRd").}

\item{paths_significant_only}{Only map the significant paths and select a variable to filter by.}

\item{paths_significant_cutoff}{Choose a p-value cutoff.}

\item{paths_weight}{Add weight to the pathway lines in the format "W10".}

\item{paths_opacity}{Set opacity of highlighted lines. Value between 0 and 1.}

\item{include_compounds}{Map compounds in addition to pathways. Can be only significant or all.}

\item{compounds_color}{Put one color in hex form here for the compounds.}

\item{compounds_weight}{Add weight to the highlighted compounds in the format "W10"}

\item{compounds_opacity}{Set opacity of highlighted compounds. Value between 0 and 1.}

\item{highlight_path_names}{Boolean to bold the highlighted KEGG pathways.}

\item{module_names_font_color}{Color of the module labels (i.e. Carbohydrate metabolism, Vitamins and co-factors).}

\item{module_names_font_weight}{Weight of the module labels. Options are 'bold', 'normal', 'bolder', 'lighter', or a numeric between 1 and 1000.}

\item{module_names_background_color}{Change the background colors of the module labels. iPath defaults to colors of the rainbow. This allows you to insert a color in hex or color names (e.g. "grey") to change all the module fill colors at the same time.}

\item{module_names_background_stroke}{This allows you to put a boarder around the module label background. Default is no stroke.}

\item{generate_legend}{Boolean generates a legend in the R graphics window.}

\item{df}{A data frame with 1-4 columns. The first column should be 'KEGG.ID' (can be compounds "C00001" or paths "map00012"), then 'Color' in hex form, then 'Line.Width' in the from "W12", then 'Opacity' between 0 and 1.}
}
\description{
This function will generate the KEGG atlas heatmap from a MetaboAnalyst Pathways enrichment analysis.
}
