% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggMapsToNames.R
\name{keggMapsToNames}
\alias{keggMapsToNames}
\title{Convert KEGG IDs to pathway names}
\usage{
keggMapsToNames(map_ids, keep_all = TRUE, return_df = FALSE)
}
\arguments{
\item{map_ids}{A vector of KEGG IDs}

\item{keep_all}{Boolean: TRUE keeps rows that were not found in the path to names data. FALSE removes these.}

\item{return_df}{Boolean: TRUE returns a data frame with 2 columns (i.e. the input vector and the matching pathway name). FALSE just returns the matching pathway vector}
}
\description{
Convert KEGG IDs to pathway names
}
