% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namesToKeggMaps.R
\name{namesToKeggMaps}
\alias{namesToKeggMaps}
\title{Convert pathway names to Kegg IDs}
\usage{
namesToKeggMaps(path_names, keep_all = TRUE, return_df = FALSE)
}
\arguments{
\item{path_names}{A vector of pathway names}

\item{keep_all}{Boolean: TRUE keeps rows that were not found in the path to names data. FALSE removes these.}

\item{return_df}{Boolean: TRUE returns a data frame with 2 columns (i.e. the input vector and the matching KEGG.IDs). FALSE just returns the matching KEGG.ID vector}
}
\description{
Convert pathway names to Kegg IDs
}
