% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathToMap data.R
\docType{data}
\name{pathToMap}
\alias{pathToMap}
\title{KEGG ID and pathway name list}
\format{
An object of class \code{data.frame} with 188 rows and 2 columns.
}
\usage{
load("data/pathToMap.Rda")
}
\description{
KEGG ID and pathway name list
}
\keyword{datasets}
