% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKeggAtlas.R
\name{getKeggAtlas}
\alias{getKeggAtlas}
\title{Generate a Kegg Atlas Heatmap}
\usage{
getKeggAtlas(
  df,
  download_file_path = getwd(),
  download_file_name = "RiPath3 Kegg Atlas.png",
  highlight_path_names = FALSE,
  module_names_font_color = "#ffffff",
  module_names_font_weight = "bold",
  module_names_background_color = "Default",
  module_names_background_stroke = "Default"
)
}
\arguments{
\item{df}{A data frame with 1-4 columns. The first column should be 'KEGG.ID' (can be compounds "C00001" or paths "map00012"), then 'Color' in hex form, then 'Line.Width' in the from "W12", then 'Opacity' between 0 and 1.}

\item{highlight_path_names}{Boolean to bold the highlighted KEGG pathways.}

\item{module_names_font_color}{Color of the module labels (i.e. Carbohydrate metabolism, Vitamins and co-factors).}

\item{module_names_font_weight}{Weight of the module labels. Options are 'bold', 'normal', 'bolder', 'lighter', or a numeric between 1 and 1000.}

\item{module_names_background_color}{Change the background colors of the module labels. iPath defaults to colors of the rainbow. This allows you to insert a color in hex or color names (e.g. "grey") to change all the module fill colors at the same time.}

\item{module_names_background_stroke}{This allows you to put a boarder around the module label background. Default is no stroke.}
}
\description{
This function runs a headless browser using 'chromote' and opens the iPath3 metabolic pathways website: 'https://pathways.embl.de/ipath3.cgi?map=metabolic'.
It will submit your data frame with pathways or compounds and allow you to do minor customizations.
It will then download a png of your data to a specified path name.
IMPORTANT: This function requires internet access and requires that you have Google Chrome installed.
}
\examples{
# has functions to make a color gradient from a numeric vector
library(scales)

# Define data frame
df <- data.frame(
  Pathway = c(
    "Arginine biosynthesis",
    "Caffeine metabolism",
    "Glycine, serine and threonine metabolism",
    "Lysine degradation",
    "Arginine and proline metabolism",
    "beta-Alanine metabolism",
    "Glutathione metabolism"),
  value = c(
    0.77393471,
    0.42070720,
    0.12953268,
    -1.74740108,
    -0.58241999,
    -1.30774442,
    -1.01095672)
  )

# Convert names to IDs
df$KEGG.ID <- namesToKeggMaps(df$Pathway)

# Give a color palette
my_cols <- brewer.pal(7, "YlOrRd") # From RColorBrewer
color_gradient <- col_numeric(palette = my_cols, domain = range(df$value))
df$color <- color_gradient(df$value)

# Add width
df$width <- "W12"

getKeggAtlas(
  df[, c("KEGG.ID", "color", "width")],
  download_file_path = getwd(),
  highlight_path_names = TRUE,
  module_names_font_color = "black",
  module_names_font_weight = "bolder",
  module_names_background_color = "#dddddd",
  module_names_background_stroke = "black"
 )
}
